AP_Mount: extract shared #TP-frame protocol layer into AP_Mount_Backend_TPFrame
AP_Mount_Topotek and AP_Mount_SkyDroid speak the same "#TP"/"#tp" wire
framing (courtesy of a shared OEM hardware supplier), but each carried its
own independently-duplicated implementation of the framing/CRC/packet-send
layer - byte-for-byte identical in several functions. Adds
AP_Mount_Backend_TPFrame, deriving from AP_Mount_Backend_Serial, which both
now derive from instead, and moves the shared parts into it:

- read_incoming_packets()'s parser state machine, calculate_crc(),
  hex2char(), send_fixedlen_packet(), send_variablelen_packet(), and the
  HeaderType/ControlByte/ParseState/Identifier types and _msg_buff/
  _msg_buff_len/_parser members.
- A handle_message(msg_id) pure-virtual hook the shared parser calls once a
  packet's CRC is verified, so each product's own command dispatch stays
  entirely its own.
- packetlen_max()/is_valid_address_byte()/source_address_byte() virtuals let
  each product plug in its own packet-size limit, valid address-byte set,
  and outgoing source-address rule (Topotek's varies with whether the port
  is network-attached; SkyDroid's protocol doc confirms 'U' covers both
  UART and UDP connections, so it doesn't need to vary).

Each product's own AddressByte enum, command-identifier set, and all
gimbal_*_analyse()/request_*()/send_target_*() functions are untouched and
stay entirely separate - only the protocol-content-free framing layer moved.
A thin same-signature wrapper in each subclass keeps every existing
send_fixedlen_packet()/send_variablelen_packet() call site unchanged.

Named after the literal "#TP"/"#tp" marker rather than either company, since
neither product's protocol document ever names or expands what "TP" stands
for, and the marker is shared by both regardless of which one originated it.

Addresses Peter Barker's review of the SkyDroid driver PR
(ArduPilot/ardupilot#34155): "framing, parser, CRC, and packet-send
functions are copy-pasted from Topotek, and both drivers being compiled in
doubles that flash cost" - confirmed directly against the code before
starting this (byte-for-byte identical in several functions), then measured
via three independent regression passes: SkyDroid's own three autotests
against the new base alone, Topotek's two existing autotests (MountTopotek,
MountTopotekNetwork) with zero regressions once Topotek moved to the same
base, and all five together as a final combined pass - all green throughout.
