+++ libraries/AP_Mount/AP_Mount_SkyDroid.cpp
@@ -14,17 +14,7 @@ extern const AP_HAL::HAL& hal;
 
 #define AP_MOUNT_SKYDROID_UPDATE_INTERVAL_MS 100                 // resend angle or rate targets, and push our attitude, at this interval
 #define AP_MOUNT_SKYDROID_HEALTH_TIMEOUT_MS  1000                // timeout for health (based on attitude reports from gimbal)
-#define AP_MOUNT_SKYDROID_PACKETLEN_MIN      12                  // packet length not including the data segment
-#define AP_MOUNT_SKYDROID_DATALEN_MAX        (AP_MOUNT_SKYDROID_PACKETLEN_MAX - AP_MOUNT_SKYDROID_PACKETLEN_MIN) // data segment len can be no more than this
 #define AP_MOUNT_SKYDROID_ATTITUDE_RATE_HZ   50                  // rate we ask the gimbal to stream its attitude to us (matches the 50hz rate AP_Mount::update() is actually called at; doc allows up to 100hz)
-
-// byte offsets within a received packet - see the packet-format table in this file's
-// header comment.  Every packet shares this same preamble layout regardless of command;
-// only the data segment's own internal layout (if any) differs per command, and is
-// documented at each parse function that reads one
-#define AP_MOUNT_SKYDROID_MSGOFS_DATALEN 5    // data length, 1 ASCII hex nibble
-#define AP_MOUNT_SKYDROID_MSGOFS_ID      7    // 3-character command identifier
-#define AP_MOUNT_SKYDROID_MSGOFS_DATA    10   // start of the command-specific data segment
 
 // 3 character identifiers
 #define AP_MOUNT_SKYDROID_ID3CHAR_GIMBAL_MODE       "PTZ"        // discrete gimbal control, data bytes: 00:stop, 01:up, 02:down, 03:left, 04:right, 05:center, 06:follow, 07:lock head.  Only the follow/lock codes (06/07 - see set_gimbal_lock()) and center (05 - see send_center_command()) are used
@@ -244,171 +234,21 @@ bool AP_Mount_SkyDroid::get_attitude_quaternion(Quaternion& att_quat)
     return true;
 }
 
-// reading incoming packets from gimbal and confirm they are of the correct format
-void AP_Mount_SkyDroid::read_incoming_packets()
+// dispatch on the 3-character command ID to the function that consumes that
+// message - called by AP_Mount_Backend_TPFrame::read_incoming_packets() once a
+// packet's CRC has been verified
+void AP_Mount_SkyDroid::handle_message(const char* msg_id)
 {
-    // check for bytes on the serial port
-    const uint16_t nbytes = MIN(_uart->available(), 1024U);
-    if (nbytes == 0) {
-        return;
-    }
-
-    // flag to allow cases below to reset parser state
-    bool reset_parser = false;
-
-    // process bytes received
-    for (uint16_t i = 0; i < nbytes; i++) {
-        uint8_t b;
-        if (!_uart->read(b)) {
-            continue;
-        }
-
-        // add latest byte to buffer
-        _msg_buff[_msg_buff_len++] = b;
-
-        // protect against overly long messages
-        if (_msg_buff_len >= AP_MOUNT_SKYDROID_PACKETLEN_MAX) {
-            reset_parser = true;
-        }
-
-        // process byte depending upon current state
-        switch (_parser.state) {
-
-        case ParseState::WAITING_FOR_HEADER1:
-            if (b == '#') {
-                _parser.state = ParseState::WAITING_FOR_HEADER2;
-                break;
-            }
-            reset_parser = true;
-            break;
-
-        case ParseState::WAITING_FOR_HEADER2:
-            // 't'/'T' (and 'p'/'P' below) distinguish HeaderType::VARIABLE_LEN from
-            // FIXED_LEN - meaningful on transmit (see send_variablelen_packet()), but
-            // deliberately not tracked here on receive: the packet's own Data_Len
-            // nibble is authoritative regardless of which header case was used to
-            // send it, so the parser has no need to remember which one it saw
-            if (b == 't' || b == 'T') {
-                _parser.state = ParseState::WAITING_FOR_HEADER3;
-                break;
-            }
-            reset_parser = true;
-            break;
-
-        case ParseState::WAITING_FOR_HEADER3:
-            if (b == 'p' || b == 'P') {
-                _parser.state = ParseState::WAITING_FOR_ADDR1;
-                break;
-            }
-            reset_parser = true;
-            break;
-
-        case ParseState::WAITING_FOR_ADDR1:
-        case ParseState::WAITING_FOR_ADDR2:
-            if (b == 'U' || b == 'M' || b == 'D' || b == 'G') {
-                // advance to next state
-                _parser.state = (ParseState)((uint8_t)_parser.state+1);
-                break;
-            }
-            reset_parser = true;
-            break;
-
-        case ParseState::WAITING_FOR_DATALEN: {
-            // sanity check data length
-            uint8_t data_len;
-            if (hex_char_to_nibble(b, data_len) && data_len <= AP_MOUNT_SKYDROID_DATALEN_MAX) {
-                _parser.data_len = data_len;
-                _parser.state = ParseState::WAITING_FOR_CONTROL;
-                break;
-            }
-            reset_parser = true;
-            break;
-        }
-
-        case ParseState::WAITING_FOR_CONTROL:
-            // r or w
-            if (b == 'r' || b == 'w') {
-                _parser.state = ParseState::WAITING_FOR_ID1;
-                break;
-            }
-            reset_parser = true;
-            break;
-
-        case ParseState::WAITING_FOR_ID1:
-        case ParseState::WAITING_FOR_ID2:
-        case ParseState::WAITING_FOR_ID3:
-            // check all uppercase letters and numbers.  eg 'GAC'
-            if ((b >= 'A' && b <= 'Z') || (b >= '0' && b <= '9')) {
-                // advance to next state
-                _parser.state = (ParseState)((uint8_t)_parser.state+1);
-                break;
-            }
-            reset_parser = true;
-            break;
-
-        case ParseState::WAITING_FOR_DATA: {
-            // normally hex numbers in char form (e.g. '0A')
-            const uint8_t data_bytes_received = _msg_buff_len - (AP_MOUNT_SKYDROID_PACKETLEN_MIN - 2);
-
-            // sanity check to protect against programming errors
-            if (data_bytes_received > AP_MOUNT_SKYDROID_DATALEN_MAX) {
-                INTERNAL_ERROR(AP_InternalError::error_t::flow_of_control);
-                reset_parser = true;
-                break;
-            }
-
-            // advance parser state once expected number of bytes have been received
-            if (data_bytes_received == _parser.data_len) {
-                _parser.state = ParseState::WAITING_FOR_CRC_LOW;
-            }
-            break;
-        }
-
-        case ParseState::WAITING_FOR_CRC_LOW:
-            _parser.state = ParseState::WAITING_FOR_CRC_HIGH;
-            break;
-
-        case ParseState::WAITING_FOR_CRC_HIGH:
-            // this is the last byte in the message so reset the parser
-            reset_parser = true;
-
-            // sanity check to protect against programming errors
-            if (_msg_buff_len < AP_MOUNT_SKYDROID_PACKETLEN_MIN) {
-                INTERNAL_ERROR(AP_InternalError::error_t::flow_of_control);
-                break;
-            }
-
-            // calculate and check CRC
-            const uint8_t crc_value = calculate_crc(_msg_buff, _msg_buff_len - 2);
-            const char crc_char1 = hex2char((crc_value >> 4) & 0x0f);
-            const char crc_char2 = hex2char((crc_value) & 0x0f);
-            if (crc_char1 != _msg_buff[_msg_buff_len - 2] || crc_char2 != _msg_buff[_msg_buff_len-1]) {
-                debug("CRC expected:%x got:%c%c", (int)crc_value, crc_char1, crc_char2);
-                break;
-            }
-
-            // CRC is OK, dispatch on the 3-character command ID to the function that
-            // consumes that message
-            const char *msg_id = (const char*)&_msg_buff[AP_MOUNT_SKYDROID_MSGOFS_ID];
-            if (strncmp(msg_id, "GAC", 3) == 0) {
-                gimbal_angle_analyse();
-            } else if (strncmp(msg_id, "REC", 3) == 0) {
-                gimbal_record_analyse();
-            } else if (strncmp(msg_id, "SDC", 3) == 0) {
-                gimbal_sdcard_analyse();
-            } else if (strncmp(msg_id, "VER", 3) == 0) {
-                gimbal_version_analyse();
-            } else if (strncmp(msg_id, "MOD", 3) == 0) {
-                gimbal_model_analyse();
-            }
-        }
-
-        // handle reset of parser
-        if (reset_parser) {
-            _parser.state = ParseState::WAITING_FOR_HEADER1;
-            _msg_buff_len = 0;
-            reset_parser = false;
-        }
+    if (strncmp(msg_id, "GAC", 3) == 0) {
+        gimbal_angle_analyse();
+    } else if (strncmp(msg_id, "REC", 3) == 0) {
+        gimbal_record_analyse();
+    } else if (strncmp(msg_id, "SDC", 3) == 0) {
+        gimbal_sdcard_analyse();
+    } else if (strncmp(msg_id, "VER", 3) == 0) {
+        gimbal_version_analyse();
+    } else if (strncmp(msg_id, "MOD", 3) == 0) {
+        gimbal_model_analyse();
     }
 }
 
@@ -612,9 +452,9 @@ void AP_Mount_SkyDroid::gimbal_angle_analyse()
 {
     // consume current angles.  data is yaw, pitch, roll in that order, each 4 hex chars, 0.01deg units
     uint32_t yaw_raw, pitch_raw, roll_raw;
-    if (!hex_chars_to_uint32((const char*)&_msg_buff[AP_MOUNT_SKYDROID_MSGOFS_DATA], 4, yaw_raw) ||
-        !hex_chars_to_uint32((const char*)&_msg_buff[AP_MOUNT_SKYDROID_MSGOFS_DATA + 4], 4, pitch_raw) ||
-        !hex_chars_to_uint32((const char*)&_msg_buff[AP_MOUNT_SKYDROID_MSGOFS_DATA + 8], 4, roll_raw)) {
+    if (!hex_chars_to_uint32((const char*)&_msg_buff[AP_MOUNT_TPFRAME_MSGOFS_DATA], 4, yaw_raw) ||
+        !hex_chars_to_uint32((const char*)&_msg_buff[AP_MOUNT_TPFRAME_MSGOFS_DATA + 4], 4, pitch_raw) ||
+        !hex_chars_to_uint32((const char*)&_msg_buff[AP_MOUNT_TPFRAME_MSGOFS_DATA + 8], 4, roll_raw)) {
         return;
     }
     const int16_t yaw_angle_cd = wrap_180_cd((int16_t)yaw_raw);
@@ -643,7 +483,7 @@ void AP_Mount_SkyDroid::gimbal_record_analyse()
 {
     // data is 2 ASCII chars ("00" or "01") - only the low digit is ever non-zero, so
     // that's the only one we need to check
-    _recording = (_msg_buff[AP_MOUNT_SKYDROID_MSGOFS_DATA + 1] == '1');
+    _recording = (_msg_buff[AP_MOUNT_TPFRAME_MSGOFS_DATA + 1] == '1');
 }
 
 // information analysis of gimbal storage card
@@ -654,8 +494,8 @@ void AP_Mount_SkyDroid::gimbal_sdcard_analyse()
     // all-zero (no card), matching this function's previous behaviour
     static const uint8_t all_zero_chars[10] = {'0','0','0','0','0','0','0','0','0','0'};
     bool all_zero = true;
-    if (_msg_buff_len >= AP_MOUNT_SKYDROID_MSGOFS_DATA + ARRAY_SIZE(all_zero_chars)) {
-        all_zero = (memcmp(&_msg_buff[AP_MOUNT_SKYDROID_MSGOFS_DATA], all_zero_chars, ARRAY_SIZE(all_zero_chars)) == 0);
+    if (_msg_buff_len >= AP_MOUNT_TPFRAME_MSGOFS_DATA + ARRAY_SIZE(all_zero_chars)) {
+        all_zero = (memcmp(&_msg_buff[AP_MOUNT_TPFRAME_MSGOFS_DATA], all_zero_chars, ARRAY_SIZE(all_zero_chars)) == 0);
     }
     _sdcard_healthy = !all_zero;
 }
@@ -664,8 +504,8 @@ void AP_Mount_SkyDroid::gimbal_sdcard_analyse()
 void AP_Mount_SkyDroid::gimbal_version_analyse()
 {
     uint8_t data_buf_len;
-    if (!hex_char_to_nibble(_msg_buff[AP_MOUNT_SKYDROID_MSGOFS_DATALEN], data_buf_len) || data_buf_len == 0 ||
-        _msg_buff[AP_MOUNT_SKYDROID_MSGOFS_DATA] != 'V') {
+    if (!hex_char_to_nibble(_msg_buff[AP_MOUNT_TPFRAME_MSGOFS_DATALEN], data_buf_len) || data_buf_len == 0 ||
+        _msg_buff[AP_MOUNT_TPFRAME_MSGOFS_DATA] != 'V') {
         return;
     }
 
@@ -674,7 +514,7 @@ void AP_Mount_SkyDroid::gimbal_version_analyse()
     uint8_t ver_count = 0;
     uint32_t ver_num = 0;
     for (uint8_t i = 1; i < data_buf_len && ver_count < ARRAY_SIZE(version); i++) {
-        const uint8_t c = _msg_buff[AP_MOUNT_SKYDROID_MSGOFS_DATA + i];
+        const uint8_t c = _msg_buff[AP_MOUNT_TPFRAME_MSGOFS_DATA + i];
         if (c == '.') {
             version[ver_count++] = ver_num;
             ver_num = 0;
@@ -709,106 +549,17 @@ void AP_Mount_SkyDroid::gimbal_version_analyse()
 void AP_Mount_SkyDroid::gimbal_model_analyse()
 {
     uint8_t data_buf_len;
-    if (!hex_char_to_nibble(_msg_buff[AP_MOUNT_SKYDROID_MSGOFS_DATALEN], data_buf_len) || data_buf_len == 0) {
+    if (!hex_char_to_nibble(_msg_buff[AP_MOUNT_TPFRAME_MSGOFS_DATALEN], data_buf_len) || data_buf_len == 0) {
         return;
     }
     memset(_model_name, 0, sizeof(_model_name));
-    memcpy(_model_name, _msg_buff + AP_MOUNT_SKYDROID_MSGOFS_DATA, MIN((uint8_t)(sizeof(_model_name)-1), data_buf_len));
+    memcpy(_model_name, _msg_buff + AP_MOUNT_TPFRAME_MSGOFS_DATA, MIN((uint8_t)(sizeof(_model_name)-1), data_buf_len));
 
     // display gimbal model name to user.  Informational only - no control decision
     // depends on it (see this driver's header comment)
     GCS_SEND_TEXT(MAV_SEVERITY_INFO, "%s model %s", send_message_prefix, _model_name);
 
     _got_model_name = true;
-}
-
-// calculate checksum
-uint8_t AP_Mount_SkyDroid::calculate_crc(const uint8_t *cmd, uint8_t len) const
-{
-    uint8_t crc = 0;
-    for (uint16_t i = 0; i<len; i++) {
-        crc += cmd[i];
-    }
-    return(crc);
-}
-
-// hexadecimal to character conversion
-uint8_t AP_Mount_SkyDroid::hex2char(uint8_t data) const
-{
-    if ((9 >= data)) {
-        return (data + '0');
-    } else {
-        return (data - 10 + 'A');
-    }
-}
-
-// send a fixed length packet
-bool AP_Mount_SkyDroid::send_fixedlen_packet(AddressByte address, const Identifier id, bool write, uint8_t value)
-{
-    uint8_t databuff[3];
-    hal.util->snprintf((char *)databuff, ARRAY_SIZE(databuff), "%02X", value);
-    return send_variablelen_packet(HeaderType::FIXED_LEN, address, id, write, databuff, ARRAY_SIZE(databuff)-1);
-}
-
-// send variable length packet
-bool AP_Mount_SkyDroid::send_variablelen_packet(HeaderType header, AddressByte address, const Identifier id, bool write, const uint8_t* databuff, uint8_t databuff_len)
-{
-    // exit immediately if not initialised
-    if (!_initialised) {
-        return false;
-    }
-
-    // calculate and sanity check packet size
-    const uint16_t packet_size = AP_MOUNT_SKYDROID_PACKETLEN_MIN + databuff_len;
-    if (packet_size > AP_MOUNT_SKYDROID_PACKETLEN_MAX) {
-        debug("send_packet data buff too large");
-        return false;
-    }
-
-    // check for sufficient space in outgoing buffer
-    if (_uart->txspace() < packet_size) {
-        debug("tx buffer full");
-        return false;
-    }
-
-    // create buffer for holding outgoing packet
-    uint8_t send_buff[packet_size];
-    uint8_t send_buff_ofs = 0;
-
-    // packet header (bytes 0 ~ 2)
-    send_buff[send_buff_ofs++] = '#';
-    send_buff[send_buff_ofs++] = (header == HeaderType::FIXED_LEN) ? 'T' : 't';
-    send_buff[send_buff_ofs++] = (header == HeaderType::FIXED_LEN) ? 'P' : 'p';
-
-    // address (bytes 3, 4).  source is always the UDP/external-control address for this gimbal
-    send_buff[send_buff_ofs++] = (uint8_t)AddressByte::UDP;
-    send_buff[send_buff_ofs++] = (uint8_t)address;
-
-    // data length (byte 5)
-    send_buff[send_buff_ofs++] = hex2char(databuff_len);
-
-    // control byte (byte 6)
-    send_buff[send_buff_ofs++] = write ? (uint8_t)ControlByte::WRITE : (uint8_t)ControlByte::READ;
-
-    // identifier (bytes 7 ~ 9)
-    send_buff[send_buff_ofs++] = id[0];
-    send_buff[send_buff_ofs++] = id[1];
-    send_buff[send_buff_ofs++] = id[2];
-
-    // data
-    if (databuff_len != 0) {
-        memcpy(&send_buff[send_buff_ofs], databuff, databuff_len);
-        send_buff_ofs += databuff_len;
-    }
-
-    // crc
-    uint8_t crc = calculate_crc(send_buff, send_buff_ofs);
-    send_buff[send_buff_ofs++] = hex2char((crc >> 4) & 0x0f);
-    send_buff[send_buff_ofs++] = hex2char(crc & 0x0f);
-
-    // send packet
-    _uart->write(send_buff, send_buff_ofs);
-    return true;
 }
 
 // set gimbal's lock vs follow mode
